get_top_peak_gene_pairs

For peaks and genes with high abundance, construct candidate pairs where peaks are in close proximity to the gene.

Usage

get_top_peak_gene_pairs(
  obj,
  gene_top = 2000,
  peak_top = 20000,
  distance = 5e+05,
  gene_assay = "RNA",
  peak_assay = "peak"
)

Arguments

obj

A Seurat object of single-cell multimodal data.

gene_top

Number of top highly expressed genes. Default to 2000.

peak_top

Number of top highly accessible peaks. Default to 20000.

distance

Distance threshold for peaks to be considered as possibly associated with genes. Default to 5e+05.

gene_assay

Name of the gene assay in Seurat object. Default to 'RNA'.

peak_assay

Name of the peak assay in Seurat object. Default to 'peak'.

Value

A data frame of candidate peak-gene pairs.

gene

gene in the peak-gene pair

peak

peak in the peak-gene pair