WLS utility function.

This function performs WLS for one gene and all peaks in its candidate pairs.

Usage

scMultiMap_WLS_by_gene(
  x_gene,
  X_peak,
  m_gene,
  M_peak,
  w_gene,
  W_peak,
  var_gene,
  Var_peak,
  gene_seq_depth,
  peak_seq_depth
)

Source

Cell-type-specific mapping of enhancer and target genes from single-cell multimodal data. Chang Su, Dongsoo Lee, Peng Jin, Jingfei Zhang;

Arguments

x_gene

A length n vector of gene counts. n denotes the number of cells.

X_peak

A n by p peak count matrix. p denotes the number of peaks.

m_gene

A length n vector of gene counts' mean across cells.

M_peak

A n by p matrix of peak counts' mean across cells for all peaks.

w_gene

A length n vector used to calculate regression weights.

W_peak

A n by p matrix used to calculate regression weights.

var_gene

A length n vector of gene counts' variance across cells.

Var_peak

A n by p matrix of peak counts' variance across cells for all peaks.

gene_seq_depth

A length n vector of sequencing depths from the RNA modality.

peak_seq_depth

A length n vector of sequencing depths from the ATAC/peak modality.

Value

A data.frame with p rows and three columns

pval

p-value

test_stat

test statistics

covar

estimated covariance

Each row correspond to scMultiMap's WLS result on one gene-peak pair.