A toy example of estimated results from scMultiMap_IRLS() on small_obj.

The source code for generating this data set is in data-raw/ on Github.

Usage

small_irls_list

Format

small_irls_list

A list of two lists generated from the 1st step of scMultiMap

gene

scMultiMap_IRLS results on genes

peak

scMultiMap_IRLS results on peaks