validate_with_assay

Validate peak-gene associations computed with single-cell multimodal data by evaluating the consistency with orthogonal assay.

Usage

validate_with_assay(res_df, sig_assay_gr, pvar = "padj", p_cutoff = 0.2)

Arguments

res_df

A data frame of peak-gene associations results, containing columns of gene, peak, and pvar (see below).

sig_assay_gr

A GRanges object for associated peak-gene pairs from other orthogonal assays, containing gene as a column in metadata

pvar

Which variable in res_df to use for claiming significant associations. Default to 'padj'.

p_cutoff

Cutoff for significance based on pvar. Default to 0.2.

Value

A named numeric vector that summarizes the overlap between res_df results and sig_assay_gr results

pval

p value from one-sided Fisher exact test

n_overlap

numebr of overlapped peak-gene pairs

oddsratio

odds ratio in the two-by-two table

ratio

overlapped pairs among all significant pairs from res_df

enr

enrichment of overlapped pairs among significant pairs compared to all background pairs